requirements.txt
Flask==2.3.3
torch>=2.2.0
torchvision>=0.17.0
timm>=0.9.7
Pillow>=10.0.0
numpy>=1.24.3
supabase>=2.0.0
gunicorn>=21.2.0